requirements.txt
dataclasses-json==0.4.2
paho-mqtt==1.5.0
